requirement.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.1
pycparser==2.20
PyMySQL==0.10.1
redis==3.5.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
threadpool==1.3.2
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
